Fix 64 to 32 bits casts warnings on msvc

+++ rzclientreconnect/test/AuthServer/Common.cpp
@@ -85,7 +85,7 @@ void sendGameConnectedAccounts(TestConnectionChannel *channel, std::vector<Accou
 	static const int MAXCOUNT_PER_PACKET = 2;
 
 	TS_GA_ACCOUNT_LIST* accountListPacket;
-	int maxCount = accounts.size() <= MAXCOUNT_PER_PACKET ? accounts.size() : MAXCOUNT_PER_PACKET;
+	int maxCount = (int)(accounts.size() <= MAXCOUNT_PER_PACKET ? accounts.size() : MAXCOUNT_PER_PACKET);
 	accountListPacket = TS_MESSAGE_WNA::create<TS_GA_ACCOUNT_LIST, TS_GA_ACCOUNT_LIST::AccountInfo>(maxCount);
 
 	auto it = accounts.begin();
